+++ lib/beamer/locations/settings_location.dart
@@ -222,6 +222,7 @@ class SettingsLocation extends BeamLocation<BeamState> {
           ),
           child: ProjectDetailPage(
             projectId: state.pathParameters['projectId']!,
+            categoryId: state.uri.queryParameters['categoryId'],
           ),
         ),